Adding Assimp to support multiple 3D formats
Now that Assimp is supported. tinyobj has been removed from the project. Also, some refactor has been made on render pass and pipeline, and now the  shaders are compiled in runtime and the shader is stored in a binary file.

+++ Ilargi/source/Platform/Vulkan/VulkanImGuiPanel.cpp
@@ -6,7 +6,7 @@
 #include "VulkanSwapchain.h"
 #include "VulkanCommandBuffer.h"
 
-#include "Utils/IlargiUI.h"
+#include "Utils/UI/IlargiUI.h"
 
 // 3rd Party headers
 #include <imgui.h>
@@ -122,7 +122,7 @@ namespace Ilargi
 			clearValues[0].color = { {0.0f, 0.0f, 0.0f, 1.0f} };
 			clearValues[1].depthStencil = { 1.0f, 0 };
 
-			renderPassInfo.clearValueCount = clearValues.size();
+			renderPassInfo.clearValueCount = static_cast<uint32_t>(clearValues.size());
 			renderPassInfo.pClearValues = clearValues.data();
 
 			VkViewport viewport{};
